+++ src/e/correcoes/e5/Exer1.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package e.exercicios;
+package e.correcoes.e5;
 
 /**
  *
