+++ src/e/correcoes/e5/Exer4.java
@@ -3,16 +3,15 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package e.exercicios;
+package e.correcoes.e5;
 
 /**
  *
  * @author user
  */
-public class Exer14 {
-    public static void main (String[] args){
-    
-    
-        
-    }
+public class Exer4 {
+  //  4.Faça um programa que imprima o seu nome e imprima em tela
+    public static void main(String[ ] args) {
+    System.out.println("Iben Enky");   
+}
 }
